Added Identity Service, integrated user authentication in auction service

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -47,11 +48,12 @@ public class AuctionController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto) {
         var  auction = _mapper.Map<Auction>(auctionDto);
-        // TODO : addd current user as seeler
-        auction.Seller = "test";
+
+        auction.Seller = User?.Identity?.Name ?? throw new InvalidOperationException("User is not authenticated");
         _context.Auctions.Add(auction);
 
         var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -75,7 +77,9 @@ public class AuctionController : ControllerBase
             return NotFound();
         }
 
-        // TODO: check seller == username
+        if (auction.Seller != User?.Identity?.Name) {
+            return Unauthorized("You are not authorized to update this auction");
+        }
 
         auction.Item.Make = auctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = auctionDto.Model ?? auction.Item.Model;
@@ -101,7 +105,9 @@ public class AuctionController : ControllerBase
             return NotFound();
         }
 
-        // TODO: check seller == username
+        if (auction.Seller != User?.Identity?.Name) {
+            return Unauthorized("You are not authorized to delete this auction");
+        }
 
         _context.Auctions.Remove(auction);
